requirements.txt
discord.py==1.3.2
dnspython==1.16.0
PyNaCl==1.3.0